backend/requirements.txt
# Django and Core
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File Storage
django-storages==1.14.2
boto3==1.34.0

# Environment and Configuration
python-decouple==3.8

# Security
django-ratelimit==4.1.0

# Utilities
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2

# Telegram Bot
python-telegram-bot==20.7

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0